requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
numpy==1.18.1
opencv-python==4.1.2.30
panda==0.3.1
paramiko==2.7.1
Pillow==7.0.0
pycparser==2.19
PyNaCl==1.3.0
pytesseract==0.3.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
